makefile: rename codegen to protogen
Now that we have renamed the generate target to crds, it makes sense to also rename to
codegen target to make it more clear that it specifically refers to protobuf codegen. Just
as with make generate, we keep the old codegen target around to use as an alias for
backward compatbility. We also remove the original protogen dockerfile as it is no longer
needed. Instead, we can simply use the base cilium-builder image.
