+++ tools/protoc-gen-go-tetragon/common/common.go
