+++ tools/protoc-gen-go-tetragon/eventchecker/multichecker.go
@@ -4,7 +4,7 @@
 package eventchecker
 
 import (
-	"github.com/cilium/tetragon/cmd/protoc-gen-go-tetragon/common"
+	"github.com/cilium/tetragon/tools/protoc-gen-go-tetragon/common"
 	"google.golang.org/protobuf/compiler/protogen"
 )
