+++ tools/protoc-gen-go-tetragon/eventchecker/yaml.go
@@ -7,7 +7,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/cilium/tetragon/cmd/protoc-gen-go-tetragon/common"
+	"github.com/cilium/tetragon/tools/protoc-gen-go-tetragon/common"
 	"github.com/iancoleman/strcase"
 	"google.golang.org/protobuf/compiler/protogen"
 )
