+++ tools/protoc-gen-go-tetragon/generate/generate.go
@@ -6,9 +6,9 @@ package generate
 import (
 	"fmt"
 
-	"github.com/cilium/tetragon/cmd/protoc-gen-go-tetragon/eventchecker"
-	"github.com/cilium/tetragon/cmd/protoc-gen-go-tetragon/helpers"
-	"github.com/cilium/tetragon/cmd/protoc-gen-go-tetragon/types"
+	"github.com/cilium/tetragon/tools/protoc-gen-go-tetragon/eventchecker"
+	"github.com/cilium/tetragon/tools/protoc-gen-go-tetragon/helpers"
+	"github.com/cilium/tetragon/tools/protoc-gen-go-tetragon/types"
 	"google.golang.org/protobuf/compiler/protogen"
 )
