+++ tools/protoc-gen-go-tetragon/main.go
@@ -4,7 +4,7 @@
 package main
 
 import (
-	"github.com/cilium/tetragon/cmd/protoc-gen-go-tetragon/generate"
+	"github.com/cilium/tetragon/tools/protoc-gen-go-tetragon/generate"
 )
 
 func main() {
