+++ tools/protoc-gen-go-tetragon/types/types.go
@@ -6,7 +6,7 @@ package types
 import (
 	"path/filepath"
 
-	"github.com/cilium/tetragon/cmd/protoc-gen-go-tetragon/common"
+	"github.com/cilium/tetragon/tools/protoc-gen-go-tetragon/common"
 	"google.golang.org/protobuf/compiler/protogen"
 )
